+++ src/sites.rs
@@ -3,7 +3,7 @@ use std::ops::Range;
 #[derive(Debug)]
 pub struct Sites {
     /// sorted genome-wide positions
-    gw_pos_cm: Vec<f32>,
+    gw_pos_cm: Vec<f64>,
     gw_pos: Vec<u32>,
     // ranges of pos for chromosome
     chrom_ranges: Vec<Range<u32>>,
@@ -17,7 +17,7 @@ impl Sites {
             chrom_ranges: vec![],
         }
     }
-    pub fn add(&mut self, gw_pos: u32, gw_pos_cm: f32) {
+    pub fn add(&mut self, gw_pos: u32, gw_pos_cm: f64) {
         self.gw_pos.push(gw_pos);
         self.gw_pos_cm.push(gw_pos_cm);
     }
@@ -56,7 +56,7 @@ impl Sites {
     pub fn get_pos_slice(&self) -> &[u32] {
         &self.gw_pos[..]
     }
-    pub fn get_pos_cm_slice(&self) -> &[f32] {
+    pub fn get_pos_cm_slice(&self) -> &[f64] {
         &self.gw_pos_cm[..]
     }
     pub fn get_chrom_pos_idx_ranges(&self, chrid: usize) -> (usize, usize) {
